Replace klab 1. Change package name 2. Change in build.gradle 3. Change in pom.xml
