+++ todo-app/build.gradle
@@ -4,7 +4,7 @@ plugins {
     id 'java'
 }
 
-group = 'com.klab.todo'
+group = 'com.psi.todo'
 version = '0.0.1-SNAPSHOT'
 sourceCompatibility = '11'
