--- todo-app/src/main/java/com/klab/todo/TodoApplication.java
@@ -1,12 +1,0 @@
-package com.klab.todo;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TodoApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(TodoApplication.class, args);
-    }
-}
