--- todo-app/src/main/java/com/klab/todo/dto/TodoDeleteRequest.java
@@ -1,14 +1,0 @@
-package com.klab.todo.dto;
-
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Positive;
-
-import lombok.Data;
-
-@Data
-public class TodoDeleteRequest {
-    @NotNull(message = "Tood id must not be null")
-    @Positive(message = "Tood id must be grater than 0")
-    private Long id;
-
-}
