--- todo-app/src/main/java/com/klab/todo/dto/TodoResource.java
@@ -1,17 +1,0 @@
-package com.klab.todo.dto;
-
-import javax.validation.constraints.NotEmpty;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonInclude;
-
-import lombok.Data;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonIgnoreProperties(ignoreUnknown = true)
-@Data
-public class TodoResource {
-    @NotEmpty(message = "Content not specified")
-    private String content;
-    private Boolean isComplete = false;
-}
