--- todo-app/src/main/java/com/klab/todo/entity/Todo.java
@@ -1,25 +1,0 @@
-package com.klab.todo.entity;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Entity
-@Table(name = "todo")
-@JsonIgnoreProperties({ "hibernateLazyInitializer", "handler" })
-public class Todo {
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-    private String content;
-    private Boolean isComplete;
-}
