--- todo-app/src/main/java/com/klab/todo/exception/TodoException.java
@@ -1,13 +1,0 @@
-package com.klab.todo.exception;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class TodoException {
-
-    private String message;
-    private Integer code;
-
-}
