--- todo-app/src/main/java/com/klab/todo/repository/TodoRepository.java
@@ -1,8 +1,0 @@
-package com.klab.todo.repository;
-
-import org.springframework.data.repository.CrudRepository;
-
-import com.klab.todo.entity.Todo;
-
-public interface TodoRepository extends CrudRepository<Todo, Long> {
-}
