--- todo-app/src/main/java/com/klab/todo/validator/Advance.java
@@ -1,4 +1,0 @@
-package com.klab.todo.validator;
-
-public interface Advance {
-}
