--- todo-app/src/main/java/com/klab/todo/validator/TodoExistsValidator.java
@@ -1,26 +1,0 @@
-package com.klab.todo.validator;
-
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
-
-import org.springframework.stereotype.Component;
-
-import com.klab.todo.repository.TodoRepository;
-
-@Component
-public class TodoExistsValidator implements ConstraintValidator<TodoExists, Long> {
-
-    private TodoRepository todoRepository;
-
-    public TodoExistsValidator(TodoRepository todoRepository) {
-        this.todoRepository = todoRepository;
-    }
-
-    public boolean isValid(Long id, ConstraintValidatorContext context) {
-        return isTodoPresent(id);
-    }
-
-    private boolean isTodoPresent(Long id) {
-        return todoRepository.findById(id).isPresent();
-    }
-}
