--- todo-app/src/main/java/com/klab/todo/validator/ValidationSequence.java
@@ -1,8 +1,0 @@
-package com.klab.todo.validator;
-
-import javax.validation.GroupSequence;
-import javax.validation.groups.Default;
-
-@GroupSequence({ Default.class, Minimal.class, Advance.class })
-public interface ValidationSequence {
-}
