+++ todo-app/src/main/java/com/psi/todo/TodoApplication.java
@@ -1,0 +1,12 @@
+package com.psi.todo;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TodoApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(TodoApplication.class, args);
+    }
+}
