+++ todo-app/src/main/java/com/psi/todo/dto/TodoResource.java
@@ -1,0 +1,17 @@
+package com.psi.todo.dto;
+
+import javax.validation.constraints.NotEmpty;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonInclude;
+
+import lombok.Data;
+
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@JsonIgnoreProperties(ignoreUnknown = true)
+@Data
+public class TodoResource {
+    @NotEmpty(message = "Content not specified")
+    private String content;
+    private Boolean isComplete = false;
+}
