+++ todo-app/src/main/java/com/psi/todo/exception/TodoException.java
@@ -1,0 +1,13 @@
+package com.psi.todo.exception;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class TodoException {
+
+    private String message;
+    private Integer code;
+
+}
