+++ todo-app/src/main/java/com/psi/todo/repository/TodoRepository.java
@@ -1,0 +1,8 @@
+package com.psi.todo.repository;
+
+import org.springframework.data.repository.CrudRepository;
+
+import com.psi.todo.entity.Todo;
+
+public interface TodoRepository extends CrudRepository<Todo, Long> {
+}
