+++ todo-app/src/main/java/com/psi/todo/validator/Advance.java
@@ -1,0 +1,4 @@
+package com.psi.todo.validator;
+
+public interface Advance {
+}
