+++ todo-app/src/main/java/com/psi/todo/validator/Minimal.java
@@ -1,0 +1,4 @@
+package com.psi.todo.validator;
+
+public interface Minimal {
+}
