+++ todo-app/src/main/java/com/psi/todo/validator/ValidationSequence.java
@@ -1,0 +1,8 @@
+package com.psi.todo.validator;
+
+import javax.validation.GroupSequence;
+import javax.validation.groups.Default;
+
+@GroupSequence({ Default.class, Minimal.class, Advance.class })
+public interface ValidationSequence {
+}
